// prisma/generated/type-graphql/models/CasePk.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../scalars";
import { Renamedcase } from "../models/Renamedcase";
import { User } from "../models/User";

@TypeGraphQL.ObjectType({
  isAbstract: true
})
export class CasePk {
  @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
    nullable: false
  })
  id!: number;

  @TypeGraphQL.Field(_type => Date, {
    nullable: false
  })
  createdAt!: Date;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  createdBy?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  didampingi?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  legalMemo?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  notulaRapat?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  targetAkhirAdvokasi?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  strategiAdvokasi?: string | null;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  statusAlasanTdk?: string | null;

  @TypeGraphQL.Field(_type => Date, {
    nullable: true
  })
  tglRapat?: Date | null;

  @TypeGraphQL.Field(_type => Date, {
    nullable: false
  })
  updatedAt!: Date;

  @TypeGraphQL.Field(_type => String, {
    nullable: true
  })
  updatedBy?: string | null;

  @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
    nullable: true
  })
  pp_pendamping?: number | null;

  ppPendamping?: User | null;

  case?: Renamedcase | null;
}